fix: 엔티티 JPA 타입 정합성 + Redis 포트 충돌 회피

+++ api/crimp-core/src/main/java/io/crimp/core/entity/log/SessionAttempt.java
@@ -44,7 +44,7 @@ public class SessionAttempt {
     private AttemptResult result;
 
     @Column(name = "attempts", nullable = false)
-    private Integer attempts;
+    private Short attempts;
 
     @Column(name = "media_id")
     private Long mediaId;
@@ -62,7 +62,7 @@ public class SessionAttempt {
         this.sessionId = sessionId;
         this.routeId = routeId;
         this.result = result;
-        this.attempts = attempts;
+        this.attempts = (short) attempts;
         this.loggedAt = loggedAt;
     }
